✨ Add loading pages
