--- apps/web/src/app/[locale]/poll/[urlId]/loading.tsx
@@ -1,3 +1,0 @@
-export default function Loading() {
-  return null;
-}
